tests/wpt/web-platform-tests/css/css-lists/css-lists-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'counter-increment',
  from: 'initial',
  to: 'add 123'
});

test_no_interpolation({
  property: 'counter-increment',
  from: 'add 5',
  to: 'add 123'
});

test_no_interpolation({
  property: 'counter-reset',
  from: 'initial',
  to: 'add 123'
});

test_no_interpolation({
  property: 'counter-set',
  from: 'initial',
  to: 'add 123'
});

test_no_interpolation({
  property: 'list-style-position',
  from: 'initial',
  to: 'inside'
});

test_no_interpolation({
  property: 'list-style-type',
  from: 'initial',
  to: 'square'
});
</script>
